requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
crispy-bootstrap5==2024.10
dj-database-url==0.5.0
Django==4.2.13
django-allauth==0.57.2
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
django-ranged-response==0.2.0
django-ratelimit==4.1.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-summernote==0.8.20.0
gunicorn==20.1.0
oauthlib==3.2.2
pillow==11.0.0
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests-oauthlib==2.0.0
sqlparse==0.5.1
stripe==11.2.0
whitenoise==6.5.0
